+++ formuledevoyage_card.php
@@ -83,8 +83,13 @@ require_once DOL_DOCUMENT_ROOT.'/core/class/html.formprojet.class.php';
 dol_include_once('/clienjoyholidays/class/formuledevoyage.class.php');
 dol_include_once('/clienjoyholidays/lib/clienjoyholidays_formuledevoyage.lib.php');
 
+// Hook for business offer
+if (isModEnabled("propal")) {
+	require_once DOL_DOCUMENT_ROOT.'/comm/propal/class/propal.class.php';
+}
+
 // Load translation files required by the page
-$langs->loadLangs(array("clienjoyholidays@clienjoyholidays", "other"));
+$langs->loadLangs(array("clienjoyholidays@clienjoyholidays", "propal", "other"));
 
 // Get parameters
 $id = GETPOST('id', 'int');