+++ ChangeLog.md
@@ -5,7 +5,7 @@
 - NEW: AJAX Request *08/08/2024* - 1.2
 
 ## 1.1
-- NEW: Default Costs depending on destination country - *05/08/2024 - 1.1*
+- NEW: Default Costs depending on destination country - *05/08/2024* - 1.1
 
 ## 1.0
 Initial version