+++ ChangeLog.md
@@ -1,8 +1,13 @@
 # CHANGELOG CLIENJOYHOLIDAYS FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
-## 1.0
 
+## 1.2
+- NEW: AJAX Request *08/08/2024* - 1.2
+
+## 1.1
+- NEW: Default Costs depending on destination country - *05/08/2024* - 1.1
+
+## 1.0
 Initial version
 
-##1.1
-- NEW: Default Costs depending on destination country - *05/08/2024* - 1.1
+